Separate JS code to where it should belong
