+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"main":[{"name":"main-5e1bc4bce4925aa0c5d4.js","path":"/sites/betterself/assets/bundles/main-5e1bc4bce4925aa0c5d4.js"}]}}
+{"status":"done","chunks":{"main":[{"name":"main-eef99e9d60504b68840f.js","path":"/sites/betterself/assets/bundles/main-eef99e9d60504b68840f.js"}]},"error":"ModuleNotFoundError","message":"Module not found: Error: Can't resolve './urls/constants' in '/sites/betterself/assets/js/auth'"}
